+++ go/proto/streams/managementpb/service.pb.go
@@ -24,7 +24,7 @@ var File_streams_management_service_proto protoreflect.FileDescriptor
 
 const file_streams_management_service_proto_rawDesc = "" +
 	"\n" +
-	" streams/management/service.proto\x12\x18freym.streams.management\x1a streams/management/publish.proto\x1a\"streams/management/subscribe.proto\x1a\x1estreams/management/event.proto\x1a\x1dstreams/management/gdpr.proto\x1a!streams/management/paginate.proto\x1a$streams/management/backchannel.proto\x1a\x1fstreams/management/stream.proto\x1a!streams/management/snapshot.proto\x1a\x1fstreams/management/rename.proto\x1a$streams/management/transaction.proto2\x98\x10\n" +
+	" streams/management/service.proto\x12\x18freym.streams.management\x1a streams/management/publish.proto\x1a\"streams/management/subscribe.proto\x1a\x1estreams/management/event.proto\x1a\x1dstreams/management/gdpr.proto\x1a!streams/management/paginate.proto\x1a$streams/management/backchannel.proto\x1a\x1fstreams/management/stream.proto\x1a!streams/management/snapshot.proto\x1a\x1fstreams/management/rename.proto\x1a$streams/management/transaction.proto\x1a\"streams/management/erroneous.proto2\xa5\x12\n" +
 	"\aService\x12^\n" +
 	"\aPublish\x12(.freym.streams.management.PublishRequest\x1a).freym.streams.management.PublishResponse\x12h\n" +
 	"\tSubscribe\x12*.freym.streams.management.SubscribeRequest\x1a+.freym.streams.management.SubscribeResponse(\x010\x01\x12V\n" +
@@ -41,7 +41,9 @@ const file_streams_management_service_proto_rawDesc = "" +
 	"\x0eInvalidateGdpr\x12/.freym.streams.management.InvalidateGdprRequest\x1a0.freym.streams.management.InvalidateGdprResponse\x12\x85\x01\n" +
 	"\x14CreateStreamSnapshot\x125.freym.streams.management.CreateStreamSnapshotRequest\x1a6.freym.streams.management.CreateStreamSnapshotResponse\x12v\n" +
 	"\x0fRenameEventType\x120.freym.streams.management.RenameEventTypeRequest\x1a1.freym.streams.management.RenameEventTypeResponse\x12\xa6\x01\n" +
-	"\x1fWaitForTransactionalConsistency\x12@.freym.streams.management.WaitForTransactionalConsistencyRequest\x1aA.freym.streams.management.WaitForTransactionalConsistencyResponse\x12j\n" +
+	"\x1fWaitForTransactionalConsistency\x12@.freym.streams.management.WaitForTransactionalConsistencyRequest\x1aA.freym.streams.management.WaitForTransactionalConsistencyResponse\x12\x82\x01\n" +
+	"\x13ListErroneousEvents\x124.freym.streams.management.ListErroneousEventsRequest\x1a5.freym.streams.management.ListErroneousEventsResponse\x12\x85\x01\n" +
+	"\x14ResendErroneousEvent\x125.freym.streams.management.ResendErroneousEventRequest\x1a6.freym.streams.management.ResendErroneousEventResponse\x12j\n" +
 	"\vBackchannel\x12,.freym.streams.management.BackchannelRequest\x1a-.freym.streams.management.BackchannelResponseb\x06proto3"
 
 var file_streams_management_service_proto_goTypes = []any{
@@ -61,21 +63,25 @@ var file_streams_management_service_proto_goTypes = []any{
 	(*CreateStreamSnapshotRequest)(nil),             // 13: freym.streams.management.CreateStreamSnapshotRequest
 	(*RenameEventTypeRequest)(nil),                  // 14: freym.streams.management.RenameEventTypeRequest
 	(*WaitForTransactionalConsistencyRequest)(nil),  // 15: freym.streams.management.WaitForTransactionalConsistencyRequest
-	(*BackchannelRequest)(nil),                      // 16: freym.streams.management.BackchannelRequest
-	(*PublishResponse)(nil),                         // 17: freym.streams.management.PublishResponse
-	(*SubscribeResponse)(nil),                       // 18: freym.streams.management.SubscribeResponse
-	(*Event)(nil),                                   // 19: freym.streams.management.Event
-	(*PaginateEventsResponse)(nil),                  // 20: freym.streams.management.PaginateEventsResponse
-	(*PaginateEventsAfterEventIdResponse)(nil),      // 21: freym.streams.management.PaginateEventsAfterEventIdResponse
-	(*IsStreamEmptyResponse)(nil),                   // 22: freym.streams.management.IsStreamEmptyResponse
-	(*PaginateStreamResponse)(nil),                  // 23: freym.streams.management.PaginateStreamResponse
-	(*PaginateStreamAfterEventIdResponse)(nil),      // 24: freym.streams.management.PaginateStreamAfterEventIdResponse
-	(*IntroduceGdprOnEventFieldResponse)(nil),       // 25: freym.streams.management.IntroduceGdprOnEventFieldResponse
-	(*InvalidateGdprResponse)(nil),                  // 26: freym.streams.management.InvalidateGdprResponse
-	(*CreateStreamSnapshotResponse)(nil),            // 27: freym.streams.management.CreateStreamSnapshotResponse
-	(*RenameEventTypeResponse)(nil),                 // 28: freym.streams.management.RenameEventTypeResponse
-	(*WaitForTransactionalConsistencyResponse)(nil), // 29: freym.streams.management.WaitForTransactionalConsistencyResponse
-	(*BackchannelResponse)(nil),                     // 30: freym.streams.management.BackchannelResponse
+	(*ListErroneousEventsRequest)(nil),              // 16: freym.streams.management.ListErroneousEventsRequest
+	(*ResendErroneousEventRequest)(nil),             // 17: freym.streams.management.ResendErroneousEventRequest
+	(*BackchannelRequest)(nil),                      // 18: freym.streams.management.BackchannelRequest
+	(*PublishResponse)(nil),                         // 19: freym.streams.management.PublishResponse
+	(*SubscribeResponse)(nil),                       // 20: freym.streams.management.SubscribeResponse
+	(*Event)(nil),                                   // 21: freym.streams.management.Event
+	(*PaginateEventsResponse)(nil),                  // 22: freym.streams.management.PaginateEventsResponse
+	(*PaginateEventsAfterEventIdResponse)(nil),      // 23: freym.streams.management.PaginateEventsAfterEventIdResponse
+	(*IsStreamEmptyResponse)(nil),                   // 24: freym.streams.management.IsStreamEmptyResponse
+	(*PaginateStreamResponse)(nil),                  // 25: freym.streams.management.PaginateStreamResponse
+	(*PaginateStreamAfterEventIdResponse)(nil),      // 26: freym.streams.management.PaginateStreamAfterEventIdResponse
+	(*IntroduceGdprOnEventFieldResponse)(nil),       // 27: freym.streams.management.IntroduceGdprOnEventFieldResponse
+	(*InvalidateGdprResponse)(nil),                  // 28: freym.streams.management.InvalidateGdprResponse
+	(*CreateStreamSnapshotResponse)(nil),            // 29: freym.streams.management.CreateStreamSnapshotResponse
+	(*RenameEventTypeResponse)(nil),                 // 30: freym.streams.management.RenameEventTypeResponse
+	(*WaitForTransactionalConsistencyResponse)(nil), // 31: freym.streams.management.WaitForTransactionalConsistencyResponse
+	(*ListErroneousEventsResponse)(nil),             // 32: freym.streams.management.ListErroneousEventsResponse
+	(*ResendErroneousEventResponse)(nil),            // 33: freym.streams.management.ResendErroneousEventResponse
+	(*BackchannelResponse)(nil),                     // 34: freym.streams.management.BackchannelResponse
 }
 var file_streams_management_service_proto_depIdxs = []int32{
 	0,  // 0: freym.streams.management.Service.Publish:input_type -> freym.streams.management.PublishRequest
@@ -94,26 +100,30 @@ var file_streams_management_service_proto_depIdxs = []int32{
 	13, // 13: freym.streams.management.Service.CreateStreamSnapshot:input_type -> freym.streams.management.CreateStreamSnapshotRequest
 	14, // 14: freym.streams.management.Service.RenameEventType:input_type -> freym.streams.management.RenameEventTypeRequest
 	15, // 15: freym.streams.management.Service.WaitForTransactionalConsistency:input_type -> freym.streams.management.WaitForTransactionalConsistencyRequest
-	16, // 16: freym.streams.management.Service.Backchannel:input_type -> freym.streams.management.BackchannelRequest
-	17, // 17: freym.streams.management.Service.Publish:output_type -> freym.streams.management.PublishResponse
-	18, // 18: freym.streams.management.Service.Subscribe:output_type -> freym.streams.management.SubscribeResponse
-	19, // 19: freym.streams.management.Service.GetEvent:output_type -> freym.streams.management.Event
-	19, // 20: freym.streams.management.Service.GetLastEvent:output_type -> freym.streams.management.Event
-	19, // 21: freym.streams.management.Service.GetLastHandledEvent:output_type -> freym.streams.management.Event
-	19, // 22: freym.streams.management.Service.GetLastEventByTypes:output_type -> freym.streams.management.Event
-	20, // 23: freym.streams.management.Service.PaginateEvents:output_type -> freym.streams.management.PaginateEventsResponse
-	21, // 24: freym.streams.management.Service.PaginateEventsAfterEventId:output_type -> freym.streams.management.PaginateEventsAfterEventIdResponse
-	22, // 25: freym.streams.management.Service.IsStreamEmpty:output_type -> freym.streams.management.IsStreamEmptyResponse
-	23, // 26: freym.streams.management.Service.PaginateStream:output_type -> freym.streams.management.PaginateStreamResponse
-	24, // 27: freym.streams.management.Service.PaginateStreamAfterEventId:output_type -> freym.streams.management.PaginateStreamAfterEventIdResponse
-	25, // 28: freym.streams.management.Service.IntroduceGdprOnEventField:output_type -> freym.streams.management.IntroduceGdprOnEventFieldResponse
-	26, // 29: freym.streams.management.Service.InvalidateGdpr:output_type -> freym.streams.management.InvalidateGdprResponse
-	27, // 30: freym.streams.management.Service.CreateStreamSnapshot:output_type -> freym.streams.management.CreateStreamSnapshotResponse
-	28, // 31: freym.streams.management.Service.RenameEventType:output_type -> freym.streams.management.RenameEventTypeResponse
-	29, // 32: freym.streams.management.Service.WaitForTransactionalConsistency:output_type -> freym.streams.management.WaitForTransactionalConsistencyResponse
-	30, // 33: freym.streams.management.Service.Backchannel:output_type -> freym.streams.management.BackchannelResponse
-	17, // [17:34] is the sub-list for method output_type
-	0,  // [0:17] is the sub-list for method input_type
+	16, // 16: freym.streams.management.Service.ListErroneousEvents:input_type -> freym.streams.management.ListErroneousEventsRequest
+	17, // 17: freym.streams.management.Service.ResendErroneousEvent:input_type -> freym.streams.management.ResendErroneousEventRequest
+	18, // 18: freym.streams.management.Service.Backchannel:input_type -> freym.streams.management.BackchannelRequest
+	19, // 19: freym.streams.management.Service.Publish:output_type -> freym.streams.management.PublishResponse
+	20, // 20: freym.streams.management.Service.Subscribe:output_type -> freym.streams.management.SubscribeResponse
+	21, // 21: freym.streams.management.Service.GetEvent:output_type -> freym.streams.management.Event
+	21, // 22: freym.streams.management.Service.GetLastEvent:output_type -> freym.streams.management.Event
+	21, // 23: freym.streams.management.Service.GetLastHandledEvent:output_type -> freym.streams.management.Event
+	21, // 24: freym.streams.management.Service.GetLastEventByTypes:output_type -> freym.streams.management.Event
+	22, // 25: freym.streams.management.Service.PaginateEvents:output_type -> freym.streams.management.PaginateEventsResponse
+	23, // 26: freym.streams.management.Service.PaginateEventsAfterEventId:output_type -> freym.streams.management.PaginateEventsAfterEventIdResponse
+	24, // 27: freym.streams.management.Service.IsStreamEmpty:output_type -> freym.streams.management.IsStreamEmptyResponse
+	25, // 28: freym.streams.management.Service.PaginateStream:output_type -> freym.streams.management.PaginateStreamResponse
+	26, // 29: freym.streams.management.Service.PaginateStreamAfterEventId:output_type -> freym.streams.management.PaginateStreamAfterEventIdResponse
+	27, // 30: freym.streams.management.Service.IntroduceGdprOnEventField:output_type -> freym.streams.management.IntroduceGdprOnEventFieldResponse
+	28, // 31: freym.streams.management.Service.InvalidateGdpr:output_type -> freym.streams.management.InvalidateGdprResponse
+	29, // 32: freym.streams.management.Service.CreateStreamSnapshot:output_type -> freym.streams.management.CreateStreamSnapshotResponse
+	30, // 33: freym.streams.management.Service.RenameEventType:output_type -> freym.streams.management.RenameEventTypeResponse
+	31, // 34: freym.streams.management.Service.WaitForTransactionalConsistency:output_type -> freym.streams.management.WaitForTransactionalConsistencyResponse
+	32, // 35: freym.streams.management.Service.ListErroneousEvents:output_type -> freym.streams.management.ListErroneousEventsResponse
+	33, // 36: freym.streams.management.Service.ResendErroneousEvent:output_type -> freym.streams.management.ResendErroneousEventResponse
+	34, // 37: freym.streams.management.Service.Backchannel:output_type -> freym.streams.management.BackchannelResponse
+	19, // [19:38] is the sub-list for method output_type
+	0,  // [0:19] is the sub-list for method input_type
 	0,  // [0:0] is the sub-list for extension type_name
 	0,  // [0:0] is the sub-list for extension extendee
 	0,  // [0:0] is the sub-list for field type_name
@@ -134,6 +144,7 @@ func file_streams_management_service_proto_init() {
 	file_streams_management_snapshot_proto_init()
 	file_streams_management_rename_proto_init()
 	file_streams_management_transaction_proto_init()
+	file_streams_management_erroneous_proto_init()
 	type x struct{}
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{